Rename `components` package to `component` in `:core:designsystem`
